CFrames use bytes now.
int32's where too big and inflated file size needlessly, this is a small commit cause the next ones gonna be big and i wanna seperate them

+++ RhythmThing/Objects/VideoPlayer.cs
@@ -85,13 +85,13 @@ namespace RhythmThing.Objects
                 //_timePassed += time;
                 if(_chart.vBeat >= _timePerFrame * _currentFrame)
                 {
-                    int[,] toLoad = null;
+                    byte[,] toLoad = null;
                     visual.localPositions.Clear();
 
                     while (_chart.vBeat >= _timePerFrame*(_currentFrame) && (_currentFrame != _frames))
                     {
                         
-                        toLoad = (int[,])formatter.Deserialize(readStream);
+                        toLoad = (byte[,])formatter.Deserialize(readStream);
 
                         _currentFrame++;
                     }
